changed userId type from server

+++ app/src/main/java/com/inovatech/smartpack/ui/screens/userConfig/ChangePasswordScreen.kt
@@ -50,11 +50,11 @@ import androidx.compose.runtime.setValue
 import kotlinx.coroutines.delay
 
 @Serializable
-data class ChangePassword(val id: Int)
+data class ChangePassword(val id: Long)
 
 @Composable
 fun ChangePasswordScreen(
-    userId: Int,
+    userId: Long,
     onBackPressed: () -> Unit,
     viewModel: ChangePasswordViewModel = hiltViewModel(),
 ) {
